+++ src/components/common/markets/market-cards/index.ts
@@ -1,4 +1,4 @@
 export * from './market-group-card'
 export * from './daily-market-timer'
-export * from './daily-market-card'
-export * from './daily-market-card-mobile'
+export * from './market-card'
+export * from './market-card-mobile'
